search/indexer/pyproject.toml
[tool.poetry]
name = "indexer"
version = "0.1.0"
description = "Indexador de processos para o desafio de estágio de busca do Justarter"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
elasticsearch = "^8.15.1"
psycopg2-binary = "^2.9.9"
logging = "^0.4.9.6"
pytest = "^8.3.3"
httpx = "^0.27.2"
asyncio = "^3.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
